Hide the git-repo form in the create project modal [INS-1900] (#9558)
* Hide the git-repo form in the create project modal when the git-sync feature is disabled, remove redundant props

* Fix type error
